[svn r4582] * Changed cache directory layout. comp-get update now stores comps.xml fragments in /var/cache/cl-tools/comps. Also, available component information is now stored in /var/cache/cl-tools/comps/available, and installed component information is now stored in /var/cache/cl-tools/comps/installed.
* Support for subcomponents. comp-get update creates links
in /var/cache/cl-tools/comps/available to the fragments in
/var/cache/cl-tools/comps for each subcomponent name, so
the user can now "comp-get install" etc. using those names.

* Use aptitude instead of apt-get, so we can use aptitude's
support for automatically installed packages (i.e. packages
pulled in to satisfy a dependency are automatically removed
when they are no longer needed).

* Always succeed on remove. This deals with the case where
a package cannot be removed because a package in another
component depends on it. We need to deal with
this case more intelligently (see comment in comp_remove).

* comp_update: Build a list of packages to install and remove
and only call aptitude install and dpkg --remove once, rather
than once per installed component.

* comp_update: Deal with the case where a package has moved
from one component to another (i.e. remove it from the list
of packages to be removed).

* General code cleanup.