[svn r4858] Remove the command: pdk create [workspace name] Add the command: pdk workspace create [workspace name] Modify existing test base to reflect above changes.
--------------------------------------------------------------
Note: the handling of workspace commands provides a good model
(and generalized support code) for handling of other pdk commands
of the form:
pdk entityname operation args

The same is true of the help support code for such commands.
--------------------------------------------------------------

Correct handling of arguments passed to pdk via the pdk shell.

Add testing of the help system.
Add testing which covers use of pdk from within the pdk shell.
(More of this would be a really good idea.)

Make a few other minor, collatoral improvements.