[svn r4864] Remove command: pdk updatechannels Add command pdk channel update (for the same purpose) Update testing to reflect this change
As part of the above, add "channel" handling to bin/pdk and
move resolve.do_updatechannels to channels.update.

Move resolve.do_resolve to component.resolve,
and make the functionality part of the ComponentDescriptor class.

Move resolve.do_download to component.download,
and make the functionality part of the ComponentDescriptor class.