[svn r5351] Find an incomplete complement for packages as they are loaded from components.
A complement is the source for a binary or the binaries for a source
package.

All we are doing at this point is preserving the relationship between
packages which already exists in a component descriptor after resolving.

The complement may be incomplete. This is by design to keep things
simple and avoid huge consumption of machine resources. The complement
will be used to match certain rules. (i.e. all binaries for a given
source, etc.)