[svn r5494] Clean up log.py.
Get rid of the ridiculous locking scheme. It is redundant with the
python import locking mechanism. Use a module local variable instead of
a function attribute, which is just a clearer way of storing the logger
instance in a global varaible.

Log at INFO level to make the logger more useful.