[svn r5501] Fix repogen for stanzas resolved from rpm-md repositories.
This narrowly dodges a brown paper bag error where I was testing for a
working resolve command for rpm-md repos, but repogen wouldn't actually
work. This patch adds a test for repogen folling rpm-md resolve and
makes the changes neccessary to make it pass. *whew*

In the process I've had to take a risk and declare that None == 0 for
rpm epochs, as createrepo stops epoch = None, recording it as 0. I don't
like this but if you have a better idea, speak up.