[svn r5514] Add Progeny SVN keyword to every file lacking it (exceptions: files with no comment syntax, XML files, and the copy of the GPL).
Move Progeny SVN keyword nearer to the top of the file in files
that already possess it.

Set svn:keywords property to "Progeny" on all files containing
the keyword so that it will actually expand.

Remove manually-included self-references to the filename in
many cases; these will now be handled by the keyword expansion,
and several were out of date.

Tidy up some comments.