[svn r5523] Update copyright notices to include the year 2006 in files that have seen substantial change this year.
(What constitutes a copyrightable change is not a matter
subject to easy objective analysis.  In the U.S., court cases
have established that a work title is not itself subject to
copyright protection, while a haiku is.  Yet the former can be
longer than the latter, especially when it comes to academic
literature.  Even this simple example suggests that "length" is
the sole dispositive criterion of copyrightability, which it is
not.

Anyway, I used the FSF's reputed "10 lines" rule, and applied
it to additions of algorithmic code or documentation.
Deletions don't count, the renaming of variables (or changes in
the way they're accessed), and restructuring of code to a
different indentation level also doesn't count.

Can I go back to doing something fun, now?)